requirements.txt
aiodns==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
Cython==3.0.8
dynaconf==3.2.4
frozenlist==1.4.1
idna==3.6
lxml==5.1.0
multidict==6.0.4
orjson==3.9.12
py-cord==2.4.1
pycares==4.4.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.2.1
requests==2.31.0
soupsieve==2.5
typing_extensions==4.9.0
urllib3==2.1.0
xmltodict==0.13.0
yarl==1.9.4
